requirements.txt
# Core Deep Learning
torch>=2.3.0
torchvision>=0.18.0
# YOLO26 requires ultralytics>=8.4.0 (released Jan 2026)
# Also supports YOLOv10, YOLOv11, YOLOv12
ultralytics>=8.4.0

# Data Augmentation
albumentations>=1.4.0

# Image Processing
opencv-python>=4.9.0
Pillow>=10.0.0

# ONNX Export
onnx>=1.16.0
onnxruntime-gpu>=1.18.0
onnx-simplifier>=0.4.0
onnxscript>=0.1.0

# Utilities
numpy>=1.26.0
tqdm>=4.66.0
pyyaml>=6.0
torchinfo>=1.8.0

# Logging & Tracking
wandb

# Text Generation (for synthetic data)
trdg
